+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/ErrorMessageFormat.java
@@ -11,7 +11,11 @@ public enum ErrorMessageFormat implements MessageFormatter {
 
     TABLE_UNDEFINED("Table name '%s' does not exist."),
     TABLE_ALIGNMENT_UNDEFINED("Table '%s' cannot be aligned to a request with intervals: %s."),
-    TABLE_SCHEMA_UNDEFINED("Table '%s' is incompatible with the dimensions '%s', metrics '%s' and granularity '%s' requested."),
+    TABLE_SCHEMA_UNDEFINED(
+            "Table '%s' is incompatible with the dimensions '%s', metrics '%s' and granularity '%s' requested.",
+            "No PhysicalTable with the availability-based dimensions '%s', metrics '%s', and granularity '%s' found " +
+                    "for Logical Table '%s'"
+    ),
 
     EMPTY_DICTIONARY("%s Dictionary is empty."),