+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Current
 
 ### Added:
 
+- [More descriptive log message when no physical table found due to schema mismatch](https://github.com/yahoo/fili/pull/113)
+    * Previous log message was the user-facing one, and not as helpful as it could have been
+
 - [Ability to not try to cache Druid responses that are larger than the maximum size supported by the cache implementation](https://github.com/yahoo/fili/pull/93)
     * Supported for both Cache v1 and V2
     * Controlled with `bard__druid_max_response_length_to_cache` setting